+++ Tests/UpdateBarMenuBarTests/DashboardSidebarViewControllerTests.swift
@@ -67,18 +67,22 @@
             _ = controller.view
             let tableView = try XCTUnwrap(findTableView(in: controller.view))
 
-            XCTAssertEqual(tableView.numberOfRows, 6)
-            XCTAssertEqual(tableView.selectedRow, 5)
+            XCTAssertEqual(tableView.numberOfRows, 7)
+            XCTAssertEqual(tableView.selectedRow, 6)
             XCTAssertEqual(
                 (controller.tableView(tableView, viewFor: nil, row: 3)?.accessibilityLabel()),
-                "Logs"
+                "Templates"
             )
             XCTAssertEqual(
                 (controller.tableView(tableView, viewFor: nil, row: 4)?.accessibilityLabel()),
-                "Settings"
+                "Logs"
             )
             XCTAssertEqual(
                 (controller.tableView(tableView, viewFor: nil, row: 5)?.accessibilityLabel()),
+                "Settings"
+            )
+            XCTAssertEqual(
+                (controller.tableView(tableView, viewFor: nil, row: 6)?.accessibilityLabel()),
                 "About"
             )
         }